refactor: Move regex-like functions to new `regex-utils` file
